geotiff tests: consolidate release-gate registry
Fold the 13 ``test_release_gate_*.py`` files (and the unmarked tests
that lived alongside them) into a single registry at
``xrspatial/geotiff/tests/release_gates/test_stable_features.py``. The
release engineer now has one audit point: ``pytest -m release_gate``
selects exactly this file and no others.

Every test keeps ``@pytest.mark.release_gate``; tests previously living
in a ``test_release_gate_*.py`` file but missing the marker (overview
sidecar metadata, rotated negative cases, the #2321 meta-gates) pick it
up here. Helper-function name collisions across the source files
(``_write_known_good`` in four files, ``_make_data_array`` in two)
are resolved with section prefixes so the consolidation does not
introduce cross-section coupling.

``docs/source/reference/release_gate_geotiff.rst`` and the smaller
``docs/source/reference/geotiff.rst`` references now point at the
consolidated module. The per-feature grouping in the doc is preserved
in prose even though the underlying file is one.

PR 10 of 11 in epic #2390. Closes #2403.
